added typeconverters and util for date formatting
