+++ app/src/main/java/com/nehamadiwal/jetnoteapp/data/NoteDatabase.kt
@@ -2,38 +2,13 @@ package com.nehamadiwal.jetnoteapp.data
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
-import androidx.room.TypeConverter
 import androidx.room.TypeConverters
 import com.nehamadiwal.jetnoteapp.model.Note
-import java.util.Date
-import java.util.UUID
+import com.nehamadiwal.jetnoteapp.util.DateConverter
+import com.nehamadiwal.jetnoteapp.util.UuidConverter
 
 @Database(entities = [Note::class], version = 1, exportSchema = false)
-@TypeConverters(DateConverter::class, UUIDConverter::class)
-abstract class NoteDatabase: RoomDatabase() {
+@TypeConverters(DateConverter::class, UuidConverter::class)
+abstract class NoteDatabase : RoomDatabase() {
     abstract fun noteDao(): NoteDatabaseDao
-}
-
-class DateConverter {
-    @TypeConverter
-    fun timeStampFromDate(date: Date): Long {
-        return date.time
-
-    }
-    @TypeConverter
-    fun dateFromTimestamp(timestamp: Long): Date? {
-        return Date(timestamp)
-    }
-}
-
-class UUIDConverter {
-    @TypeConverter
-    fun fromUUID(uuid: UUID): String? {
-        return uuid.toString()
-    }
-
-    @TypeConverter
-    fun uuidFromString(string: String?): UUID? {
-        return UUID.fromString(string)
-    }
 }
