+++ app/src/main/java/com/nehamadiwal/jetnoteapp/screen/NoteViewModel.kt
@@ -36,12 +36,12 @@ class NoteViewModel @Inject constructor(private val repository: NoteRepository) 
         }
     }
 
-    suspend fun addNote(note: Note) = viewModelScope.launch { repository.addNote(note = note) }
+    fun addNote(note: Note) = viewModelScope.launch { repository.addNote(note = note) }
 
-    suspend fun updateNote(note: Note) =
+    fun updateNote(note: Note) =
         viewModelScope.launch { repository.updateNote(note = note) }
 
-    suspend fun removeNote(note: Note) =
+    fun removeNote(note: Note) =
         viewModelScope.launch { repository.deleteNote(note = note) }
