+++ app/src/main/java/com/nehamadiwal/jetnoteapp/util/DateConverter.kt
@@ -1,0 +1,16 @@
+package com.nehamadiwal.jetnoteapp.util
+
+import androidx.room.TypeConverter
+import java.util.Date
+
+class DateConverter {
+    @TypeConverter
+    fun timeStampFromDate(date: Date): Long {
+        return date.time
+    }
+
+    @TypeConverter
+    fun dateFromTimestamp(timestamp: Long): Date? {
+        return Date(timestamp)
+    }
+}
