+++ app/src/main/java/com/nehamadiwal/jetnoteapp/util/UuidConverter.kt
@@ -1,0 +1,16 @@
+package com.nehamadiwal.jetnoteapp.util
+
+import androidx.room.TypeConverter
+import java.util.UUID
+
+class UuidConverter {
+    @TypeConverter
+    fun fromUuid(uuid: UUID): String? {
+        return uuid.toString()
+    }
+
+    @TypeConverter
+    fun uuidFromString(string: String): UUID? {
+        return UUID.fromString(string)
+    }
+}
