+++ app/src/main/java/com/nehamadiwal/jetnoteapp/util/utils.kt
@@ -1,0 +1,11 @@
+package com.nehamadiwal.jetnoteapp.util
+
+import java.text.SimpleDateFormat
+import java.util.Date
+import java.util.Locale
+
+fun formatDate(time: Long): String {
+    val date = Date(time)
+    val format = SimpleDateFormat("EEE, d, MMM hh:mm aaa", Locale.getDefault())
+    return format.format(date)
+}
